LPS-55041 Move PortletKey entry to the module
