+++ modules/apps/workflow/workflow-definition-link-web/src/com/liferay/workflow/definition/link/web/portlet/WorkflowDefinitionLinkPortlet.java
@@ -15,7 +15,7 @@
 package com.liferay.workflow.definition.link.web.portlet;
 
 import com.liferay.portal.kernel.portlet.bridges.mvc.MVCPortlet;
-import com.liferay.portal.util.PortletKeys;
+import com.liferay.workflow.definition.link.web.portlet.constants.WorkflowDefinitionLinkPortletKeys;
 
 import javax.portlet.Portlet;
 
@@ -41,7 +41,8 @@ import org.osgi.service.component.annotations.Component;
 		"javax.portlet.expiration-cache=0",
 		"javax.portlet.init-param.template-path=/",
 		"javax.portlet.init-param.view-template=/view.jsp",
-		"javax.portlet.name=" + PortletKeys.WORKFLOW_DEFINITION_LINK,
+		"javax.portlet.name=" +
+			WorkflowDefinitionLinkPortletKeys.WORKFLOW_DEFINITION_LINK,
 		"javax.portlet.resource-bundle=content.Language",
 		"javax.portlet.security-role-ref=power-user,user",
 		"javax.portlet.supports.mime-type=text/html"
